apps/backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "zergo-backend"
version = "0.1.0"
description = "ZERGO QR FastAPI backend"
authors = [{ name = "ZERGO" }]
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.110.0",
  "uvicorn[standard]>=0.23.0",
  "pydantic>=2.5.0",
  "pydantic-settings>=2.1.0",
  "httpx>=0.27.0",
  "python-dotenv>=1.0.1",
  "structlog>=24.1.0",
  "sentry-sdk[fastapi]>=2.13.0",
  "supabase>=2.6.0",
  "PyJWT[crypto]>=2.8.0",
  "pytest>=8.3.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]
